% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCPgSD_BF.R
\name{gMCPgSD_BF}
\alias{gMCPgSD_BF}
\title{Performs a graph based multiple test procedure using Bonferroni test for a given graph and
unadjusted p-values}
\usage{
gMCPgSD_BF(g, w.start, t, h, p, alpha = 0.025, timing, sf,
  sfpar = NULL, debug)
}
\arguments{
\item{g}{Matrix of local alpha to be passed. Each element defined how much of the local alpha reserved for the hypothesis corresponding to its row index is passed on to the hypothesis corresponding to its column index.}

\item{w.start}{Vector of initial weights}

\item{t}{Stage index}

\item{h}{Enpoint index}

\item{p}{Vector of unadjusted p-values}

\item{alpha}{A numeric specifying the maximal allowed type one error rate}

\item{timing}{Interim analysis timing}

\item{sf}{Spending function name}

\item{sfpar}{Spending function parameter}

\item{debug}{Debug indicator}
}
\value{
Vector indicating whether each endpoint can be rejected
}
\description{
Performs a graph based multiple test procedure using Bonferroni test for a given graph and
unadjusted p-values
}
\examples{

}
